# Replace name and entity regular expressions with specific functions for ~15% performance improvement