@jacktuck 2017-06-28
LGTM - more readable too :)

Nice work @lovell!